backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
efficientnet_pytorch
opencv-python
numpy
Pillow
httpx
supabase
python-dotenv
pydantic

# ── Advanced Forensic Pipeline ──
scipy
scikit-image
networkx
neo4j

# ── GraphRAG & LLM Reasoning ──
llama-index
langchain
langchain-community
